+++ functions/constants/responseMessage.js
@@ -2,6 +2,7 @@ module.exports = {
   NULL_VALUE: '필요한 값이 없습니다',
   OUT_OF_VALUE: '파라미터 값이 잘못되었습니다',
   NO_AUTHENTICATED: '접근 권한이 없습니다',
+  DB_ERROR: '데이터베이스 오류',
   INTERNAL_SERVER_ERROR: '서버 내 오류',
 
   // 회원가입
