+++ functions/service/scheduleService.js
@@ -1,66 +1,4 @@
-const db = require('../db/db');
-const { scheduleDao, groupDB, stickerDB } = require('../db');
-
-const util = require('../lib/util');
-const statusCode = require('../constants/statusCode');
-const responseMessage = require('../constants/responseMessage');
-const { dateCalcurator, termCalcurator } = require('../lib/dateCalcurater');
-
-/**
- * addSchedule
- * 약 추가 및 스케줄 생성 서비스
- * @param pillId 해당 약 아이디
- * @param userId 해당 약을 추가하는 유저 아이디
- * @param takeInterval 복용 간격 선택 ex) 매일, 특정 요일, 특정 간격
- * @param day 특정 요일 ex) 월, 수, 금
- * @param specific 특정 간격 ex) 1week
- * @param time 시간 리스트
- * @param start 복용 시작 날짜
- * @param end 복용 종료 날짜
- * @returns 추가된 약의 pillName
- */
-const addSchedule = async (pillId, userId, takeInterval, day, specific, time, start, end) => {
-  let client;
-  const log = `scheduleDao.addSchedule | pillId = ${pillId}, userId = ${userId}, takeInterval = ${takeInterval}, day = ${day}, specific = ${specific}, time = ${time}, start = ${start}, end = ${end}`;
-
-  try {
-    client = await db.connect(log);
-
-    // 총 날짜 수 계산
-    const term = termCalcurator(start, end);
-    // 약이 추가될 날짜 계산
-    const dateList = dateCalcurator(term, start, takeInterval, day, specific);
-
-    // 스케줄 추가 쿼리 실행
-    for (let date = 0; date < dateList.length; date++) {
-      for (let timeList = 0; timeList < time.length; timeList++) {
-        let newSchedule = await scheduleDao.addSchedule(
-          client,
-          pillId,
-          userId,
-          takeInterval,
-          dateList[date],
-          day,
-          specific,
-          time[timeList],
-          start,
-          end,
-        );
-        if (!newSchedule) throw new Error('addSchedule 스케줄 추가 중 오류 발생');
-      }
-    }
-
-    return util.success(statusCode.CREATED, responseMessage.PILL_ADDITION_SUCCESS);
-  } catch (error) {
-    console.error('addSchedule error 발생: ' + error);
-  } finally {
-    client.release();
-  }
-};
-
 module.exports = {
-  addSchedule,
-
   /**
   getMyCalendar: async (user, date) => {
     let client;
